requirements.txt
streamlit>=1.28.0
sqlite3
pandas>=2.0.0
plotly>=5.15.0
google-generativeai>=0.3.0
geopy>=2.3.0
requests>=2.28.0
python-dotenv>=1.0.0
numpy>=1.24.0